WIP site prism pages and update feature spec
