+++ spec/features/contacts/contacts_spec.rb
@@ -59,12 +59,14 @@ feature "Contacts address book", js: true do
   scenario "branston user can view addresses and create a new address" do
     click_on "Addresses"
     then_expect_heading_to_read_address_book
+    and_expect_contact_type_details_to_be_present
     and_expect_contact_details_to_be_present
 
     click_on "Add new address"
-    and_fill_in_and_submit_new_contact_details
-    and_expect_to_see_success_message
-    then_expect_new_details_to_be_present
+    and_fill_in_and_submit_new_contact_type_details
+    # and_fill_in_and_submit_new_contact_address_details
+    # and_expect_to_see_success_message
+    # then_expect_new_details_to_be_present
   end
 
   scenario "user can edit an existing address" do
@@ -337,7 +339,6 @@ feature "Contacts address book", js: true do
       postcode: "AF6 9JO",
       data_request_name: "Sue Jones",
       data_request_emails: "sue.jones@gmail.com",
-      contact_type: "probation",
     }
 
     contacts_edit_page.edit_contact(details)
@@ -353,18 +354,31 @@ feature "Contacts address book", js: true do
     expect(page).to have_content("FE2 9JK")
   end
 
-  def and_fill_in_and_submit_new_contact_details
+  def and_expect_contact_type_details_to_be_present
+    expect(page).to have_content("Prison")
+    expect(page).to have_content("Probation centre")
+  end
+
+  def and_fill_in_and_submit_new_contact_address_details
     details = {
       name: "John's law",
       address_line_1: "345 some road",
       postcode: "FG9 5IK",
       data_request_name: "John Smith",
       data_request_emails: "john.smith@gmail.com",
-      contact_type: "solicitor",
     }
 
     contacts_new_page.new_contact(details)
     click_on "Submit"
+  end
+
+  def and_fill_in_and_submit_new_contact_type_details
+    details = {
+      contact_type: "solicitor",
+    }
+
+    contacts_new_page.new_contact(details)
+    click_on "Continue"
   end
 
   def and_expect_to_see_success_message
